break phyloTree into files II
